Ami_Coding_Pari_Na/requirements.txt
Django==3.2.5
gunicorn==20.0.4
djangorestframework==3.14.0